Address syscall search review feedback ##search
